config/urls.py
"""Root URL configuration for hackthon-backend."""
from __future__ import annotations

from django.contrib import admin
from django.urls import include, path

urlpatterns = [
    path("admin/", admin.site.urls),
    path("api/", include("apps.insights.urls")),
]
